requirements.txt
Flask==2.1.3                 
authlib==1.2.1
python-dotenv==0.21.1        
requests==2.28.2             
pytest==7.2.2                
Flask-Cors==3.0.10 
transformers==4.30.2         
torch==2.1.0    
Werkzeug<2.1.0
 numpy~=1.26.4

